+++ core/src/main/java/io/questdb/cutlass/http/HttpAuthenticator.java
@@ -24,13 +24,9 @@
 
 package io.questdb.cutlass.http;
 
-import io.questdb.cairo.SecurityContext;
-import io.questdb.std.Mutable;
-import io.questdb.std.ObjList;
-import io.questdb.std.QuietCloseable;
-import org.jetbrains.annotations.Nullable;
+import io.questdb.cutlass.auth.AuthenticatorBase;
 
-public interface HttpAuthenticator extends QuietCloseable, Mutable {
+public interface HttpAuthenticator extends AuthenticatorBase {
 
     /**
      * Authenticates incoming HTTP request.
@@ -39,27 +35,4 @@ public interface HttpAuthenticator extends QuietCloseable, Mutable {
      * @return true if the authentication succeeded, false - otherwise
      */
     boolean authenticate(HttpRequestHeader headers);
-
-    @Override
-    default void clear() {
-    }
-
-    @Override
-    default void close() {
-    }
-
-    default byte getAuthType() {
-        return SecurityContext.AUTH_TYPE_NONE;
-    }
-
-    /**
-     * Returns list of groups provided by external identity provider, such as OpenID Connect provider.
-     * For other authentication types returns null.
-     */
-    @Nullable
-    default ObjList<CharSequence> getGroups() {
-        return null;
-    }
-
-    CharSequence getPrincipal();
 }
